backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "app"
version = "0.1.0"
description = "Wedge Golf Backend API"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
  {name = "Wedge Golf Team", email = "info@wedgegolf.com"}
]
dependencies = [
  "fastapi==0.104.1",
  "uvicorn==0.23.2",
  "firebase-admin==6.2.0",
  "python-multipart==0.0.6",
  "psutil==5.9.6",
  "pydantic==2.4.2",
  "requests==2.31.0",
  "pyyaml==6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest==7.4.3",
  "httpx==0.25.1",
  "email-validator==2.1.0",
]

[tool.setuptools]
package-dir = {"" = "app"}

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88 